fix: createElement 함수 단순화 및 2주차 과제를 위한 render함수 분리
